# Conflicts:
#	source/fishstudy-android/fish-study-dynamic/.idea/caches/build_file_checksums.ser
#	source/fishstudy-android/fish-study-static/.idea/caches/build_file_checksums.ser
#	source/fishstudy-android/fish-study-static/.idea/misc.xml
#	source/fishstudy-android/fish-study-static/app/src/main/java/hci201/se1171/oceanstudy/MainActivity.java
#	source/fishstudy-android/fish-study-static/app/src/main/res/layout/start_activity.xml

+++ doc/design/database/OceanStudy_Script_Table_MYSQL.sql
@@ -3,9 +3,9 @@ drop table if exists os_fish;
 create table os_fish (
 	id int auto_increment not null,
     name varchar(50),
-    weight int,
-    length int,
-    height int,
+    weight double,
+    length double,
+    height double,
     deep int,
     age int,
     img varchar(250),
